improvements to stats command
